In progress...

+++ src/main/java/com/example/demo/UserController.java
@@ -3,7 +3,11 @@ package com.example.demo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
+
 
 import java.util.Optional;
 
@@ -22,13 +26,36 @@ public class UserController {
     @PostMapping()
     public ResponseEntity<User> addNewUser(@RequestBody User user){
         User newUser = new User();
-        newUser.setId(user.getId());
+        newUser.setId((int) (userRepository.count() + 1));
         newUser.setFirstName(user.getFirstName());
         newUser.setLastName(user.getLastName());
         newUser.setEmail(user.getEmail());
         newUser.setPassword(user.getPassword());
         userRepository.save(newUser);
         return new ResponseEntity<>(newUser, HttpStatus.CREATED);
+    }
+
+
+    @PostMapping("/login")
+    public ResponseEntity<?> authenticateUser(@RequestBody User loginRequest) {
+        try {
+            System.out.println(loginRequest.getEmail());
+            User user = userRepository.findByEmail(loginRequest.getEmail())
+                    .orElseThrow(() -> new RuntimeException("User not found"));
+
+            if (loginRequest.getPassword().equals(user.getPassword())) {
+                // Authentication logic
+                return ResponseEntity.ok("Login successful");
+            } else {
+                return ResponseEntity.badRequest().body("Invalid credentials");
+            }
+        } catch (Exception e) {
+            // Log the exception
+            e.printStackTrace();
+
+            // Return an error response to the client
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Internal Server Error");
+        }
     }
 
     @GetMapping("/byId/{id}")
@@ -42,9 +69,14 @@ public class UserController {
     }
 
     @DeleteMapping("/{id}")
-    public @ResponseBody String deleteUser(@PathVariable Integer id){
-        userRepository.deleteById(id);
-        return "User deleted successfully!";
+    public @ResponseBody String deleteUser(@PathVariable Integer id, @RequestParam Integer userId){
+        User user = userRepository.findById(userId).orElseThrow(() -> new RuntimeException("User not found"));
+        if(user.getAdminFlag() == 1) {
+            userRepository.deleteById(id);
+            return "User deleted successfully!";
+        }
+        else
+            return "Not an admin.";
     }
 
 }
